backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.12
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
motor==3.7.0
mypy_extensions==1.1.0
openai==1.86.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
sniffio==1.3.1
spotipy==2.25.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
